refactor(providers): isolate OpenAI destination classification (split S02 L1/4)
